[#ADR-020] test(entitlement-preview): move full-mode rack specs into full/
The "Entitlement preview through the rack stack (ADR-020)" block boots the
app in full mode and drives /api/organizations, /api/account/permissions and
/api/colonel/entitlement-preview through the real stack. In
spec/integration/all it also ran under AUTHENTICATION_MODE=disabled, where the
before(:all) full-mode reboot can't take effect (boot! is guarded), so those
endpoints 404 and the T3 Disabled Mode job failed with 6 failures.

Relocate the block to spec/integration/full/entitlement_preview_rack_spec.rb —
its correct home, matching auth_mode_spec and the other mode-manipulating
rack specs. The mode-agnostic ColonelAPI::Logic block (direct app, mocked
strategy) stays in all/. No test logic changed.
